Implement/Derive (de)serialization instances needed for test file (#11)
# Description

Closes tweag/cardano-conformance-testing-of-consensus#52

This PR creates a separate module to accommodate the (orphan) instances
needed to serialize/deserialize a `GenesisTest` as JSON; we take this
approach to avoid spreading changes throughout the codebase during this
phase of development.

As specified in our design, `conformance-test-runner` consumes a test
file that specifies (at least) a `PointSchedule` to run.
It turns out our peer simulator would need access to the whole of a
`GenesisTest` data, which will contain the `PointSchedule` in
particular.

In #3, the relevant instances were introduced for the
`PointSchedule`alone. All this changes are moved into the aforementioned
module.