Add deforestBlockTree (#19)
# Description

An important in-progress task is to make more test code
block-polymorphic; right now it is largely hardcoded to use test blocks
but this will not work for running conformance tests against live
implementations of the protocol.

I do not know the motivation behind making these tests monomorphic;
regardless it seems the only thing the tests require is the ability to
detect when a block is on the trunk. For the conformance tests at least
we can still do this because we have a model of the entire chain.

This PR adds a helper function to make this easier by constructing a map
from each block hash in the tree to the anchored fragment from it to the
anchor. That fragment is a prefix of the trunk if and only if the block
is on the trunk.

Writing out this description it occurs to me that it might be more
efficient to make a set of all the trunk's block hashes and just test if
hashes are in the set. Though I suppose checking the entire prefix is a
stronger condition.

To run the new tests:
```
cabal test ouroboros-consensus-diffusion:consensus-test
```

---------

Co-authored-by: Nathan Bloomfield <nathan.bloomfield@tweag.io>